requirements.txt
# ============================================================
# COGNITIVE BREACH - Production Dependencies
# ============================================================
# Optimized for Streamlit Cloud (Linux environment)
# ============================================================

# Core Framework
streamlit>=1.28.0

# AI/ML
google-generativeai>=0.3.0
google-genai>=1.0.0  # Gemini SDK with TTS support

# Data Validation & Schemas
pydantic>=2.0.0

# Configuration
python-dotenv>=1.0.0

# Image Processing
Pillow>=10.0.0

# Visualization
plotly>=5.18.0

# Audio Processing (optional for voice features)
SpeechRecognition>=3.10.0

# ============================================================
# NOTES:
# - No Windows-specific packages (e.g., pywin32)
# - No local development tools (pytest, black, mypy)
# - Streamlit Cloud uses Python 3.9+ on Linux
# ============================================================
